module-ified trip planning service
feat: Implement trip refinement service and utility functions

- Added tripRefinementService.ts to handle the refinement of trip legs, optimizing walking segments after jeepney rides.
- Introduced getCoordinatesInSearchWindow function to assist in finding candidate boarding and alighting points within a specified search radius.
- Created tripUtils.ts with trimJeepLeg function to trim jeepney routes based on new start and end coordinates, ensuring accurate route representation.
