+++ modules/map/constants/tripPlanningConstants.ts
@@ -1,0 +1,21 @@
+
+export const MAX_WALK_TO_JEEP_METERS = 700;
+export const MAX_FINAL_WALK_METERS = 1500;
+export const ALIGNMENT_PROXIMITY_THRESHOLD_METERS = 200;
+export const MIN_JEEP_RIDE_PROGRESS_METERS = 200;
+export const DESIRED_PATH_SEARCH_AHEAD_METERS = 600;
+export const DIVERGENCE_THRESHOLD_METERS = 250;
+export const JEEP_TERMINAL_NO_BOARD_VERTEX_COUNT = 25;
+
+// Constants for Post-Processing Refinement Optimization
+export const REFINE_SEARCH_RADIUS_METERS = 120;
+export const REFINE_MIN_WALK_SAVING_METERS = 50;
+export const REFINE_MAX_WALK_FOR_CONNECTION = 400;
+export const MIN_CANDIDATE_SEPARATION_METERS = 40;
+export const REFINE_EARLY_EXIT_WALK_THRESHOLD = 15;
+export const ALIGHTING_POINT_SEARCH_TOLERANCE_METERS = 50;
+export const OVERLAP_CUT_THRESHOLD_METERS = 30;
+export const OVERLAP_WALK_POINTS_TO_CHECK_V7 = 4;
+export const MAX_TRIP_OPTIONS_TO_RETURN = 3;
+
+export const MAX_TRIP_LEGS = 10;
